pyproject.toml
[tool.poetry]
name = "toi-translation"
version = "0.1.0"
description = ""
authors = ["Samyak Bardiya <<>>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.110.1"
loguru = "^0.7.2"
pydantic = "^2.6.4"
pydantic-settings = "^2.2.1"
uvicorn = {extras = ["standard"], version = "^0.29.0"}

google-generativeai = "^0.4.1"
ipython = "^8.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
